Move validators to utils
Need to fix the interpolation
